refactor(delivery): break import cycle by extracting shared helpers to _derive.py
Resolves the eleven CodeQL "Module-level cyclic import" and "Cyclic import"
alerts flagged by GitHub Advanced Security on PR #2057. All eleven are the
same architectural smell projected across multiple file/function locations:
``report.py`` re-exported the renderer entry points at the bottom of the
file, and the renderer modules imported helpers back from ``report.py`` —
a cycle that worked at runtime via Python's import-time ordering but
tripped static analysis.

Resolution:

- New ``app/delivery/publish_findings/formatters/_derive.py`` owns the
  derivation and channel-syntax helpers: ``_sanitize_for_slack``,
  ``_to_telegram_html_body``, ``_mrkdwn_section``,
  ``_derive_root_cause_sentence``, ``_get_top_error_log``,
  ``_resolve_evidence_tags``, ``_format_provenance_lines``,
  ``_format_correlation_lines``, ``render_cloudwatch_link``,
  ``render_cloudwatch_link_html``, plus their private support
  (_first_sentence, _is_speculative, _remove_speculative_words,
  _star_pairs_to_bold_placeholders, _extract_log_message,
  _EVIDENCE_LOG_KEYS, _SLACK_LINK_RE).
- ``sections.py``, ``renderers/slack.py``, and ``renderers/telegram.py``
  now import from ``_derive.py`` directly. The lazy imports inside
  ``sections.py`` builder functions become top-of-file imports
  (no more cycle workaround needed).
- ``report.py`` shrinks from 481 lines to 42 — a thin re-export shim
  that re-exports the public entry points from renderers/ and the
  cloudwatch helpers from _derive.py for backwards compatibility with
  node.py and any external callers. No more bottom-of-file imports,
  no more noqa: E402.
- Dropped five dead helpers from the old report.py that no caller
  still references: ``_render_claim_lines``,
  ``_render_claim_lines_telegram``, ``_telegram_baseline_repeats_header``,
  ``_severity_telegram_header``, ``_norm_banner_key``. Their behaviour
  was replaced by section-pipeline equivalents in sections.py and
  renderers/* during the earlier refactor commits.

Import graph is now a DAG:

  _derive.py     ← leaf (only imports base, report_context, urls.aws)
  sections.py    ← _derive
  renderers/*    ← _derive, sections, severity
  report.py      ← _derive, renderers/*  (leaf re-export shim)
  node.py        ← report.py

6800 unit tests pass; lint/format/typecheck green.
