+++ app/delivery/publish_findings/formatters/_derive.py
@@ -1,0 +1,367 @@
+"""Shared derivation and channel-syntax helpers for the RCA report formatters.
+
+This module owns the pure helpers that every channel-specific renderer
+(Slack mrkdwn, Slack Block Kit, Telegram HTML, Discord markdown) and the
+:mod:`sections` builder share. Moving them out of ``report.py`` breaks the
+module-level import cycle CodeQL flagged on PR #2057: ``report.py`` is now
+a thin re-export shim that no other module depends on transitively.
+
+Three concern groups, in order of appearance below:
+
+1. **Channel-syntax conversion** — :func:`_sanitize_for_slack`,
+   :func:`_to_telegram_html_body`, :func:`_mrkdwn_section`. These translate
+   LLM-emitted Markdown into channel-native syntax (or wrap text in Slack
+   Block Kit section blocks).
+2. **Evidence / provenance / correlation extraction** —
+   :func:`_format_provenance_lines`, :func:`_format_correlation_lines`,
+   :func:`_resolve_evidence_tags`, :func:`_get_top_error_log`. These read
+   :class:`ReportContext` and return structured bullets / strings ready
+   for rendering.
+3. **Root-cause derivation** — :func:`_derive_root_cause_sentence` and its
+   private helpers. Picks the best one-sentence summary from the
+   investigation's raw root-cause text and validated claims.
+
+The CloudWatch-link helpers :func:`render_cloudwatch_link` and
+:func:`render_cloudwatch_link_html` are channel-specific renderers but
+live here too because they're called from multiple per-channel modules
+and read ctx directly; folding them into :class:`Section`'s ``extras``
+contract is a follow-up.
+
+All helpers are intentionally underscore-prefixed in the public API
+(except the two ``render_cloudwatch_link*`` functions kept for backwards
+compatibility with their original names) — they are module-internal
+implementation details consumed by ``sections.py`` and ``renderers/``.
+"""
+
+from __future__ import annotations
+
+import html
+import re
+from typing import Any
+
+from app.delivery.publish_findings.formatters.base import format_html_link, format_slack_link
+from app.delivery.publish_findings.report_context import ReportContext
+from app.delivery.publish_findings.urls.aws import build_cloudwatch_url
+
+# ---------------------------------------------------------------------------
+# Channel-syntax conversion
+# ---------------------------------------------------------------------------
+
+
+def _sanitize_for_slack(text: str) -> str:
+    """Convert standard Markdown formatting into Slack mrkdwn.
+
+    Slack does not render ``#`` headers, ``**bold**``, or other CommonMark
+    constructs. This converts the common patterns the LLM emits so the
+    rendered Slack message uses native mrkdwn (``*bold*`` and ``*headings*``).
+    """
+    result = re.sub(r"^#{1,6}\s+(.+)$", r"*\1*", text, flags=re.MULTILINE)
+    result = re.sub(r"\*\*(.+?)\*\*", r"*\1*", result)
+    return result
+
+
+_SLACK_LINK_RE = re.compile(r"<(https?://[^|>]+)(?:\|([^>]+))?>")
+
+
+def _star_pairs_to_bold_placeholders(line: str, bold_ph: dict[str, str]) -> str:
+    """Replace only paired ``*inner*`` spans (inner has no ``*``); lone ``*`` stay literal."""
+    out = line
+    while True:
+        m = re.search(r"\*([^*\n]+)\*", out)
+        if not m:
+            break
+        tok = f"«B{len(bold_ph)}»"
+        bold_ph[tok] = "<b>" + html.escape(m.group(1)) + "</b>"
+        out = out[: m.start()] + tok + out[m.end() :]
+    return out
+
+
+def _to_telegram_html_body(text: str) -> str:
+    """Convert mixed Slack-style text (headers, ``*bold*``, ``\\`code\\```, ``<url|label>``) to Telegram HTML."""  # noqa: E501
+    placeholders: dict[str, str] = {}
+
+    def _put(chunk: str) -> str:
+        token = f"«{len(placeholders)}»"
+        placeholders[token] = chunk
+        return token
+
+    s = text
+    s = re.sub(r"`([^`]+)`", lambda m: _put("<code>" + html.escape(m.group(1)) + "</code>"), s)
+    s = _SLACK_LINK_RE.sub(
+        lambda m: _put(format_html_link(m.group(2) or m.group(1), m.group(1))),
+        s,
+    )
+
+    out_lines: list[str] = []
+    for line in s.splitlines():
+        hdr = re.match(r"^#{1,6}\s+(.+)$", line)
+        if hdr:
+            out_lines.append("<b>" + html.escape(hdr.group(1).strip()) + "</b>")
+            continue
+        bold_ph: dict[str, str] = {}
+        starred = _star_pairs_to_bold_placeholders(line, bold_ph)
+        escaped = html.escape(starred)
+        for token, inner in sorted(bold_ph.items(), key=lambda kv: -len(kv[0])):
+            escaped = escaped.replace(token, inner)
+        out_lines.append(escaped)
+
+    merged = "\n".join(out_lines)
+    for token, chunk in sorted(placeholders.items(), key=lambda kv: -len(kv[0])):
+        merged = merged.replace(token, chunk)
+    return merged
+
+
+def _mrkdwn_section(text: str) -> dict[str, Any] | None:
+    """Build a Slack Block Kit section block with sanitized mrkdwn text.
+
+    Slack section blocks have a 3000 char limit per text field — we
+    truncate at 2990 to leave room for the trailing ellipsis. Returns
+    ``None`` when text is empty so callers can skip empty results.
+    """
+    sanitized = _sanitize_for_slack(text).strip()
+    if not sanitized:
+        return None
+    if len(sanitized) > 2990:
+        sanitized = sanitized[:2987] + "..."
+    return {
+        "type": "section",
+        "text": {"type": "mrkdwn", "text": sanitized},
+    }
+
+
+# ---------------------------------------------------------------------------
+# Evidence, provenance, and correlation extraction
+# ---------------------------------------------------------------------------
+
+
+def _format_provenance_lines(ctx: ReportContext) -> list[str]:
+    """Return ``• Label: summary`` bullets from ``ctx["source_provenance"]``."""
+    provenance = ctx.get("source_provenance") or {}
+    lines: list[str] = []
+    for source_name, entry in provenance.items():
+        label = entry.get("label") or source_name.title()
+        summary = entry.get("summary") or ""
+        if summary:
+            lines.append(f"• {label}: {summary}")
+    return lines
+
+
+def _format_correlation_lines(ctx: ReportContext) -> tuple[list[str], list[str]]:
+    """Return ``(signal_lines, driver_lines)`` from ``ctx["correlation"]``.
+
+    Mirrors the format introduced in PR #1877: signals show source + score,
+    drivers show confidence + rationale.
+    """
+    correlation = ctx.get("correlation") or {}
+    if not isinstance(correlation, dict):
+        return [], []
+
+    raw_signals = correlation.get("correlated_signals") or []
+    raw_drivers = correlation.get("most_likely_causal_drivers") or []
+
+    signal_lines: list[str] = []
+    for signal in raw_signals:
+        if not isinstance(signal, dict):
+            continue
+        name = signal.get("name") or "unknown"
+        source = signal.get("source") or "unknown"
+        score = signal.get("score")
+        score_text = f" score={float(score):.2f}" if isinstance(score, int | float) else ""
+        signal_lines.append(f"• {name} ({source}{score_text})")
+
+    driver_lines: list[str] = []
+    for driver in raw_drivers:
+        if not isinstance(driver, dict):
+            continue
+        name = driver.get("name") or "unknown"
+        confidence = driver.get("confidence")
+        rationale = driver.get("rationale") or ""
+        confidence_text = (
+            f" confidence={float(confidence):.2f}" if isinstance(confidence, int | float) else ""
+        )
+        suffix = f" — {_sanitize_for_slack(str(rationale))}" if rationale else ""
+        driver_lines.append(f"• {name}{confidence_text}{suffix}")
+
+    return signal_lines, driver_lines
+
+
+# Maps LLM source name → ordered list of evidence dict keys to try for a log message.
+_EVIDENCE_LOG_KEYS: dict[str, list[str]] = {
+    "datadog_logs": ["datadog_error_logs", "datadog_logs"],
+    "datadog": ["datadog_error_logs", "datadog_logs"],
+    "grafana_logs": ["grafana_error_logs", "grafana_logs"],
+    "grafana": ["grafana_error_logs", "grafana_logs"],
+    "cloudwatch_logs": ["cloudwatch_logs"],
+    "cloudwatch": ["cloudwatch_logs"],
+}
+
+
+def _extract_log_message(entry: object) -> str:
+    """Extract a plain string message from a log entry that may be a dict or a string."""
+    if isinstance(entry, dict):
+        return (entry.get("message") or "").strip()
+    return str(entry).strip()
+
+
+def _resolve_evidence_tags(text: str, evidence: dict) -> str:
+    """Replace ``[evidence: source]`` tags with the actual log message in a code span.
+
+    Tries error logs first, then all logs for the named source. If no message
+    is found the tag is removed silently to avoid leaking raw LLM annotations.
+    """
+
+    def _replace(m: re.Match) -> str:
+        source = m.group(1).strip().lower()
+        for key in _EVIDENCE_LOG_KEYS.get(source, []):
+            logs = evidence.get(key) or []
+            if logs:
+                msg = _extract_log_message(logs[0])
+                if msg:
+                    return f": `{msg}`"
+        return ""
+
+    return re.sub(r"\s*\[(?i:evidence):\s*([^\]]+)\]", _replace, text).strip()
+
+
+def _get_top_error_log(evidence: dict) -> str | None:
+    """Return the first error log message from available evidence sources."""
+    for key in (
+        "datadog_error_logs",
+        "datadog_logs",
+        "grafana_error_logs",
+        "grafana_logs",
+        "cloudwatch_logs",
+    ):
+        logs = evidence.get(key) or []
+        if logs:
+            msg = _extract_log_message(logs[0])
+            if msg:
+                return msg
+    return None
+
+
+# ---------------------------------------------------------------------------
+# Root-cause derivation
+# ---------------------------------------------------------------------------
+
+
+def _first_sentence(text: str) -> str:
+    """Return the first sentence from text, normalized to one line."""
+    cleaned = re.sub(r"(?:^|\s)#{1,6}\s+", " ", text, flags=re.MULTILINE)
+    cleaned = re.sub(
+        r"\b(?:Problem Statement|Summary|Context|Description|Overview)\b[:\s]*",
+        "",
+        cleaned,
+        flags=re.IGNORECASE,
+    )
+    normalized = " ".join(cleaned.split()).strip()
+    if not normalized:
+        return ""
+
+    parts = re.split(r"(?<=[.?!])\s+", normalized, maxsplit=1)
+    sentence = parts[0]
+    sentence = sentence.rstrip(".?!")
+    return sentence
+
+
+def _is_speculative(text: str) -> bool:
+    speculative_terms = (" may ", " might ", " possibly", " possible ", " likely ")
+    lower = f" {text.lower()} "
+    return any(term in lower for term in speculative_terms)
+
+
+def _remove_speculative_words(text: str) -> str:
+    speculative = ("may", "might", "likely", "probably", "possibly")
+    words = text.split()
+    filtered = [w for w in words if w.lower() not in speculative]
+    return " ".join(filtered)
+
+
+def _derive_root_cause_sentence(ctx: ReportContext) -> str:
+    """Derive a concise, single-sentence root cause with causal preference."""
+    root_cause_text = ctx.get("root_cause", "") or ""
+    root_cause_text = re.sub(r"\s*\[(?i:evidence):[^\]]*\]", "", root_cause_text).strip()
+    validated_claims = ctx.get("validated_claims", [])
+
+    if root_cause_text:
+        sentence = _first_sentence(root_cause_text)
+        if sentence and not _is_speculative(sentence):
+            return sentence
+
+    causal_connectors = (
+        " because ",
+        " due to ",
+        " caused ",
+        " resulted in ",
+        " led to ",
+        " root cause ",
+        " failure triggered ",
+    )
+
+    for claim_data in validated_claims:
+        claim = claim_data.get("claim", "") or ""
+        claim = re.sub(r"\s*\[(?i:evidence):[^\]]*\]", "", claim).strip()
+        lower = f" {claim.lower()} "
+        if any(connector in lower for connector in causal_connectors):
+            sentence = _first_sentence(claim)
+            if sentence:
+                return _first_sentence(_remove_speculative_words(sentence))
+
+    if root_cause_text:
+        sentence = _first_sentence(root_cause_text)
+        if sentence:
+            return sentence
+
+    if validated_claims:
+        claim = validated_claims[0].get("claim", "") or ""
+        claim = re.sub(r"\s*\[(?i:evidence):[^\]]*\]", "", claim).strip()
+        sentence = _first_sentence(claim)
+        if sentence:
+            return sentence
+
+    return ""
+
+
+# ---------------------------------------------------------------------------
+# CloudWatch link rendering (read ctx directly; channel-specific output)
+# ---------------------------------------------------------------------------
+
+
+def render_cloudwatch_link(ctx: ReportContext) -> str:
+    """Render CloudWatch logs link as Slack mrkdwn, or empty string."""
+    cw_url = ctx.get("cloudwatch_logs_url")
+    cw_group = ctx.get("cloudwatch_log_group")
+    cw_stream = ctx.get("cloudwatch_log_stream")
+
+    if cw_url:
+        return f"\n*{format_slack_link('CloudWatch Logs', cw_url)}*\n"
+    elif cw_group and cw_stream:
+        url = build_cloudwatch_url(ctx)
+        view_link = format_slack_link("CloudWatch Logs", url) if url else None
+        if view_link:
+            return f"\n*{view_link}*\n"
+        return f"\n*CloudWatch Logs:*\n* Log Group: {cw_group}\n* Log Stream: {cw_stream}\n"
+
+    return ""
+
+
+def render_cloudwatch_link_html(ctx: ReportContext) -> str:
+    """Telegram-HTML CloudWatch deep link, mirroring :func:`render_cloudwatch_link`."""
+    cw_url = ctx.get("cloudwatch_logs_url")
+    cw_group = ctx.get("cloudwatch_log_group")
+    cw_stream = ctx.get("cloudwatch_log_stream")
+
+    if cw_url:
+        safe = html.escape(str(cw_url), quote=True)
+        return f'\n<b>CloudWatch</b>: <a href="{safe}">View logs</a>\n'
+    if cw_group and cw_stream:
+        url = build_cloudwatch_url(ctx)
+        if url:
+            safe = html.escape(str(url), quote=True)
+            return f'\n<b>CloudWatch</b>: <a href="{safe}">View logs</a>\n'
+        return (
+            f"\n<b>CloudWatch Logs</b>\n"
+            f"Log Group: {html.escape(str(cw_group))}\n"
+            f"Log Stream: {html.escape(str(cw_stream))}\n"
+        )
+    return ""
